perf(db): Use u16 for Meet::num_unique_lifters
The number of lifters in a single meet is unlikely to exceed 65535.

Storing as u16 allows rustc to pack the u16 and bool into a single u32,
decreasing the size of Meet by four bytes.

+++ crates/db/src/data.rs
@@ -65,7 +65,7 @@ pub struct Meet {
     /// Number of unique competitors, by LifterID.
     /// Calculated at load-time.
     #[serde(default)]
-    pub num_unique_lifters: u32,
+    pub num_unique_lifters: u16,
 }
 
 /// The definition of an Entry in the database.
